requirements.txt
streamlit>=1.32.0
langchain-openai>=0.0.5
langchain==0.1.4
langchain-community==0.0.16
python-dotenv>=1.0.0
pypdf>=3.0.0
faiss-cpu>=1.7.4